hadd fixes

+++ hadding/hadd.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 
-
 SCRIPT_DIR="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" && pwd)"
 echo "DEBUG: Running .sh inside $SCRIPT_DIR"
 
@@ -42,11 +41,14 @@ DEST="${ALL_SPILL_DIR}/histos.root"
 echo "Input files:"
 echo "${FILES}"
 
-time root -l -b -q "${SCRIPT_DIR}/fileCheck.C(\"${DEST}\")" | grep "FILE_OK"
+if [ -e "${DEST}" ]; then
 
-if [ $? -ne 0 ]; then
-         echo "${DEST} is corrupt or zombie, skipping"
-         rm -f "${DEST}"
+  time root -l -b -q "${SCRIPT_DIR}/fileCheck.C(\"${DEST}\")" | grep "FILE_OK"
+
+  if [ $? -ne 0 ]; then
+           echo "${DEST} is corrupt or zombie, skipping"
+           rm -f "${DEST}"
+  fi
 fi
 
 for CURRENT_FILE in ${FILES}; do
